[Banner Plugin] Move test file to a new folder
